// client/src/utils/type.ts
export interface SignupData {
  email: string;
  fullName: string;
  password: string;
  cPassword: string;
  showP: boolean;
  showCP: boolean;
  imageUrl: string;
}